+++ include/utils/misc.hpp
@@ -80,6 +80,8 @@ namespace nwo5::utils {
         }
     }
 
+    cocos2d::CCRect rectOverlap(const cocos2d::CCRect& pA, const cocos2d::CCRect& pB);
+
     // very unnecessary template shenanigans but fun :33333
 
     template<typename T>